+++ Menu.h
@@ -1,7 +1,9 @@
+// CLASS: Menu
 //
-// Created by Dipesh Asd on 2022-03-10.
+// Author: Dipesh Shah, 7882947
 //
-
+// REMARKS: Stores the Menu table given to us
+//
 #pragma once
 #include <string>
 
@@ -10,7 +12,7 @@ using namespace std;
 class Menu {
 
 public:
-    float getFoodPrice(string mealName,float numIngredients);
+    float getFoodPrice(string mealName,int numIngredients);
   int getPreparationTime(string mealName,int numIngredients);
 };
